+++ lib/structure.rb
@@ -1,15 +1,18 @@
 # frozen_string_literal: true
 
-require 'pp'
+require "pp"
 
 module Structure
   module Type
     class Error < ::StandardError; end
+
     class SizeError < Error; end
+
     class MatchError < Error; end
 
     class Single
       attr_reader :classes
+
       def initialize(classes)
         @classes = classes
       end
@@ -67,7 +70,7 @@ module Structure
       end
 
       def elements_at_most
-        raise 'Wrong use of at_most' unless @number
+        raise "Wrong use of at_most" unless @number
 
         @max = @number
         @number = nil
@@ -75,7 +78,7 @@ module Structure
       end
 
       def elements_at_least
-        raise 'Wrong use of at_least' unless @number
+        raise "Wrong use of at_least" unless @number
 
         @min = @number
         @number = nil
@@ -83,7 +86,12 @@ module Structure
       end
 
       def matches?(json)
-        raise SizeError, "Size Error: #{inspect} size (#{json.size}) is not between #{@min} and #{@max}." unless json.size.between?(@min, @max)
+        unless json.size.between?(
+          @min, @max
+        )
+          raise SizeError,
+                "Size Error: #{inspect} size (#{json.size}) is not between #{@min} and #{@max}."
+        end
 
         json.all? { |j| classes.any? { |s| yield s, j } }
       end
@@ -97,6 +105,36 @@ module Structure
       end
     end
 
+    class ArrayWithStruct < Array
+      attr_reader :min, :max, :struct
+
+      def initialize(struct)
+        super([])
+        @struct = struct
+        @max = 999_999
+        @min = 1
+      end
+
+      def exactly(number)
+        with(number)
+      end
+
+      def times
+        elements
+      end
+
+      def matches?(json)
+        result = json.count do |j|
+          yield @struct, j
+        rescue Structure::Type::Error
+          false
+        end
+        raise MatchError, "#{json}\n has no #{@struct}" unless result && result.between?(@min, @max)
+
+        true
+      end
+    end
+
     module Methods
       def a_list_of(*class_list)
         Structure::Type::Array.new(class_list)
@@ -104,6 +142,10 @@ module Structure
 
       def one_of(*class_list)
         Structure::Type::Single.new(class_list)
+      end
+
+      def a_list_with(struct)
+        Structure::Type::ArrayWithStruct.new(struct)
       end
     end
   end