+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## [Unreleased]
 
+
+## [0.2.0] - 2022-03-23
+
+- Added more thorough testing
+- Added a_list_with matcher
+
+
 ## [0.1.0] - 2022-03-22
 
 - Initial release